(refactor): use separate objects to generate averages for web & mobile
