+++ spec/lib/crowdmap_averages_spec.rb
